Refactor a bit to unify with existing code
